requirements.txt
crewai
fastapi
uvicorn
requests
beautifulsoup4
serpapi
python-dotenv
google-genai
